requirements.txt
semantic-kernel
semantic-kernel[mcp]
mcp
openai
python-dotenv
gradio